Split pytest in multiple instances
