+++ .travis.yml
@@ -36,7 +36,9 @@ install:
 before_script:
 - mkdir $HOME/tmp
 - export TMPDIR=$HOME/tmp
-script: py.test _pytest --cov rasa_nlu --pep8 -v
+script:
+- py.test __pytest/base --cov rasa_nlu --pep8 -v --cov-append
+- py.test __pytest/training --cov rasa_nlu --pep8 -v --cov-append
 after_success:
 - coveralls
 deploy:
