+++ __pytest/base/test_sanity.py
@@ -1,15 +1,14 @@
-from __future__ import unicode_literals
-from __future__ import print_function
-from __future__ import division
 from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import importlib
-import logging
 import pkgutil
 from collections import defaultdict
+from multiprocessing import Queue, Process
 
 import pytest
-from multiprocessing import Queue, Process
 from six import PY2
