+++ __pytest/base/test_server.py
@@ -15,8 +15,8 @@ from rasa_nlu.config import RasaNLUConfig
 import json
 import io
 
-import utilities
-from utilities import ResponseTest
+from __pytest import utilities
+from __pytest.utilities import ResponseTest
 from rasa_nlu.server import RasaNLU
